fix(desktop): break circular import in Paywall module
Move GATED_FEATURES constant and GatedFeature type from usePaywall.ts
into constants.ts, eliminating the usePaywall → Paywall → constants →
usePaywall circular dependency. Update consumers to import from the
barrel instead of directly from usePaywall.
