+++ apps/desktop/src/renderer/components/Paywall/index.ts
@@ -1,3 +1,4 @@
+export type { GatedFeature } from "./constants";
+export { GATED_FEATURES } from "./constants";
 export { Paywall, paywall } from "./Paywall";
-export type { GatedFeature } from "./usePaywall";
-export { GATED_FEATURES, usePaywall } from "./usePaywall";
+export { usePaywall } from "./usePaywall";
