+++ apps/desktop/src/renderer/components/Paywall/usePaywall.ts
@@ -1,17 +1,8 @@
 import { authClient } from "renderer/lib/auth-client";
+import type { GatedFeature } from "./constants";
 import { paywall } from "./Paywall";
 
 type UserPlan = "free" | "pro";
-
-export const GATED_FEATURES = {
-	INVITE_MEMBERS: "invite-members",
-	INTEGRATIONS: "integrations",
-	TASKS: "tasks",
-	CLOUD_WORKSPACES: "cloud-workspaces",
-	MOBILE_APP: "mobile-app",
-} as const;
-
-export type GatedFeature = (typeof GATED_FEATURES)[keyof typeof GATED_FEATURES];
 
 export function usePaywall() {
 	const { data: session } = authClient.useSession();
